Begin to split up into modules
*Created ui and util modules
*Inside util are dagchat and rpc sub modules
*There current structure is still vastly incomplete and there are many redundant import statements
