+++ src/messages.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::util::constants::{DATA_DIR_PATH, MESSAGES_DIR_PATH};
 use chrono::prelude::DateTime;
 use chrono::{Local, NaiveDateTime, Utc};
 use rand::RngCore;
