+++ src/util/mod.rs
@@ -1,0 +1,5 @@
+pub mod coin;
+pub mod constants;
+pub mod rpc;
+pub mod userdata;
+pub mod wallet;
