+++ src/wallets.rs
@@ -1,106 +1,10 @@
 use super::*;
+use crate::util::constants::{DATA_DIR_PATH, MESSAGES_DIR_PATH, WALLETS_PATH};
+use ::bincode;
 use cursive::event::{Event, EventResult, EventTrigger, MouseEvent};
-
 use rand::RngCore;
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct WalletsAndLookup {
-    wallets_bytes: Vec<u8>,
-    lookup_bytes: Vec<u8>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Wallet {
-    pub name: String,
-    pub mnemonic: String,
-    pub seed: [u8; 32],
-    pub indexes: Vec<u32>,
-    #[serde(skip)]
-    pub accounts: Vec<Account>,
-    #[serde(skip)]
-    pub acc_idx: usize,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Account {
-    pub index: u32,
-    pub private_key: [u8; 32],
-    pub public_key: [u8; 32],
-    pub address: String,
-    pub balance: u128,
-    pub receivables: Vec<Receivable>,
-    pub messages: Result<Vec<SavedMessage>, String>,
-}
-
-impl Account {
-    fn with_index(wallet: &Wallet, index: u32, prefix: &str) -> Account {
-        let (private_key, public_key) = Account::get_keypair(&wallet.seed, index);
-        Account {
-            index,
-            private_key,
-            public_key,
-            address: get_address(&public_key, Some(prefix)),
-            balance: 0,
-            receivables: vec![],
-            messages: Ok(vec![]),
-        }
-    }
-
-    fn get_keypair(seed: &[u8; 32], index: u32) -> ([u8; 32], [u8; 32]) {
-        let private_key = dcutil::get_private_key(seed, index);
-        let public_key = Account::get_public_key(&private_key);
-        (private_key, public_key)
-    }
-    fn get_public_key(private_key: &[u8; 32]) -> [u8; 32] {
-        let dalek = ed25519_dalek::SecretKey::from_bytes(private_key).unwrap();
-        let public_key = ed25519_dalek::PublicKey::from(&dalek);
-        public_key.to_bytes()
-    }
-}
-
-impl Wallet {
-    fn new(mnemonic: String, seed: [u8; 32], name: String, prefix: &str) -> Wallet {
-        let mut wallet = Wallet {
-            name,
-            mnemonic,
-            seed,
-            indexes: vec![0],
-            accounts: vec![],
-            acc_idx: 0,
-        };
-        wallet
-            .accounts
-            .push(Account::with_index(&wallet, 0, prefix));
-        wallet
-    }
-    fn new_key(private_key: [u8; 32], name: String, prefix: &str) -> Wallet {
-        let public_key = Wallet::get_public_key(&private_key);
-        let mut wallet = Wallet {
-            name,
-            mnemonic: String::from(""),
-            seed: [0u8; 32],
-            indexes: vec![0],
-            accounts: vec![],
-            acc_idx: 0,
-        };
-        wallet.accounts.push(Account {
-            index: 0,
-            private_key,
-            public_key,
-            address: get_address(&public_key, Some(prefix)),
-            balance: 0,
-            receivables: vec![],
-            messages: Ok(vec![]),
-        });
-        wallet
-    }
-
-    fn get_public_key(private_key: &[u8; 32]) -> [u8; 32] {
-        let dalek = ed25519_dalek::SecretKey::from_bytes(private_key).unwrap();
-        let public_key = ed25519_dalek::PublicKey::from(&dalek);
-        public_key.to_bytes()
-    }
-}
+use crate::util::wallet::*;
 
 fn add_account(s: &mut Cursive, index: Option<u32>, prefix: &str) {
     let data = &mut s.user_data::<UserData>().unwrap();
